added session authentication to logging in and working on posts and comments

+++ controllers/api/comment-routes.js
@@ -1,5 +1,6 @@
 const router = require('express').Router();
 const { Comment } = require('../../models');
+const auth = require('../../utils/auth');
 
 // get all comments
 router.get('/', (req, res) => {
@@ -12,37 +13,41 @@ router.get('/', (req, res) => {
 });
 
 // create a comment
-router.post('/', (req, res) => {
-    Comment.create({
-        comment_text: req.body.comment_text,
-        user_id: req.body.user_id,
-        post_id: req.body.post_id
-    })
-    .then(dbCommentData => res.json(dbCommentData))
-    .catch(err => {
-        console.log(err);
-        res.status(500).json(err);
-    });
+router.post('/', auth, (req, res) => {
+    if (req.session) {
+        Comment.create({
+            comment_text: req.body.comment_text,
+            user_id: req.session.user_id,
+            post_id: req.body.post_id
+        })
+        .then(dbCommentData => res.json(dbCommentData))
+        .catch(err => {
+            console.log(err);
+            res.status(500).json(err);
+        });
+    }
 });
 
 // delete a comment
-router.delete('/:id', (req, res) => {
-    Comment.destroy({
-        where: {
-            id: req.params.id
-        }
-    })
-    .then(dbCommentData => {
-        if(!dbCommentData) {
-            res.status(404).json({ message: 'No comment found with this id' });
-            return;
-        }
-        res.json(dbCommentData);
-    })
-    .catch(err => {
-        console.log(err);
-        res.status(500).json(err);
-    });
+router.delete('/:id', auth, (req, res) => {
+    if(req.session) {
+        Comment.destroy({
+            where: {
+                id: req.params.id
+            }
+        })
+        .then(dbCommentData => {
+            if(!dbCommentData) {
+                res.status(404).json({ message: 'No comment found with this id' });
+                return;
+            }
+            res.json(dbCommentData);
+        })
+        .catch(err => {
+            console.log(err);
+            res.status(500).json(err);
+        });
+    }
 });
 
 module.exports = router;
